Move DonationDialog to promo dir
