+++ src/index.js
@@ -28,8 +28,6 @@ export { default as withTheme } from '@material-ui/core/styles/withTheme'
 export { ThemeProvider } from './styles'
 export { colours }
 export { default as ArticleDonationBanner } from './promo/ArticleDonationBanner'
-export { default as DonationBanner } from './promo/DonationBanner'
-export { default as ThinDonationBanner } from './promo/ThinDonationBanner'
 export { default as Autocomplete } from './Autocomplete'
 export { default as Avatar } from './Avatar'
 export { default as Button } from './Button'
@@ -45,12 +43,14 @@ export { default as DialogAvatar } from './dialog/DialogAvatar'
 export { default as DialogContent } from './dialog/DialogContent'
 export { default as DialogInlineTitle } from './dialog/DialogInlineTitle'
 export { default as DialogTitle } from './dialog/DialogTitle'
-export { default as DonationDialog } from './dialog/DonationDialog'
+export { default as DonationBanner } from './promo/DonationBanner'
+export { default as DonationDialog } from './promo/DonationDialog'
 export { default as Dropdown } from './Dropdown'
 export { default as MiniDivider } from './divider/MiniDivider'
 export { default as Person } from './Person'
 export { default as Swatch } from './Swatch'
 export { default as Switch } from './Switch'
 export { default as TextField } from './TextField'
+export { default as ThinDonationBanner } from './promo/ThinDonationBanner'
 export { default as TimePicker } from './pickers/TimePicker'
 export { default as Typography } from './Typography'
